be_python/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
librosa==0.10.1
numpy==1.24.3
pandas==2.0.3
qdrant-client==1.6.0
pydantic==2.4.2
scipy==1.11.3
matplotlib==3.7.3
python-jose==3.3.0
aiofiles==23.2.1
soundfile==0.12.1